requirements.txt
fastapi
uvicorn
python-decouple
python-multipart
pyngrok
email_reply_parser